--- a/bin/.jruby.module_opts
+++ b/bin/.jruby.module_opts
@@ -1,4 +1,4 @@
 --add-opens java.base/java.io=org.jruby.dist
 --add-opens java.base/java.nio.channels=org.jruby.dist
 --add-opens java.base/sun.nio.ch=org.jruby.dist
-
+--add-opens java.management/sun.management=org.jruby.dist