@enebo 2019-11-06
For fun here is 9.2.7.0 before some work went into unpack:

```text
Warming up --------------------------------------
  unpack1 m len: 288   577.000  i/100ms
  unpack1 m len: 576   384.000  i/100ms
 unpack1 m len: 1152   257.000  i/100ms
 unpack1 m len: 2304   146.000  i/100ms
 unpack1 m len: 4608    78.000  i/100ms
Calculating -------------------------------------
  unpack1 m len: 288      6.557k (± 3.4%) i/s -     32.889k in   5.022262s
  unpack1 m len: 576      4.345k (± 3.7%) i/s -     21.888k in   5.045440s
 unpack1 m len: 1152      2.651k (± 2.6%) i/s -     13.364k in   5.045537s
 unpack1 m len: 2304      1.467k (± 2.4%) i/s -      7.446k in   5.078940s
 unpack1 m len: 4608    775.922  (± 2.4%) i/s -      3.900k in   5.029380s
system ~/work/jruby (9.2.7.0) 1405% jruby ../snippets/unpack1.rb 
Warming up --------------------------------------
  unpack1 m len: 288   605.000  i/100ms
   unpack1 m len: 16     1.108k i/100ms
   unpack1 m len: 24     1.074k i/100ms
  unpack1 m len: 148   767.000  i/100ms
  unpack1 m len: 288   598.000  i/100ms
Calculating -------------------------------------
  unpack1 m len: 288      5.986k (± 2.1%) i/s -     30.250k in   5.056054s
   unpack1 m len: 16     11.218k (± 2.1%) i/s -     56.508k in   5.039890s
   unpack1 m len: 24     10.705k (± 6.5%) i/s -     53.700k in   5.042451s
  unpack1 m len: 148      7.779k (± 3.8%) i/s -     39.117k in   5.036783s
  unpack1 m len: 288      6.053k (± 1.7%) i/s -     30.498k in   5.039550s
```